requirements.txt
# Updated dependencies for modern Python compatibility
# Original setup.py used very old versions (2017)
# These are updated compatible versions

tensorflow>=2.10.0
keras>=2.10.0
editdistance>=0.6.0
h5py>=3.7.0
matplotlib>=3.5.0
numpy>=1.23.0
python-dateutil>=2.8.0
scipy>=1.9.0
Pillow>=9.0.0
nltk>=3.7
sk-video>=1.1.10
opencv-python>=4.6.0
deepgram-sdk
deepgram-captions